refactor: 주소 케밥 케이스, 로고 img로 변경
