+++ src/routes/auth/sign-in/index.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router';
 
-export const Route = createFileRoute('/auth/signIn/')({
+export const Route = createFileRoute('/auth/sign-in/')({
   component: RouteComponent,
 });
